add support for session logout at IdP
When a logout URL is configured, the user is redirected to that URL
after logout. We provide all the data suggested by OIDC, so if this is a
proper end_session_endpoint the user will be logged out at the IdP and
be redirected to the wiki.

When your IdP configures redirect URIs separately for authentication and
logout, be sure to configure the same redirect URI for both.